rename `MaxReportSampleRows` to `MaxSampleRows`, fix int-validation

+++ go/cmd/vtctldclient/command/vreplication/vdiff/vdiff.go
@@ -60,7 +60,7 @@ var (
 		Limit                       uint32 // We only accept positive values but pass on an int64
 		FilteredReplicationWaitTime time.Duration
 		DebugQuery                  bool
-		MaxReportSampleRows         uint64
+		MaxReportSampleRows         uint32 // We only accept positive values but pass on an int64
 		OnlyPKs                     bool
 		UpdateTableStats            bool
 		MaxExtraRowsToCompare       uint32 // We only accept positive values but pass on an int64
@@ -865,7 +865,7 @@ func registerCommands(root *cobra.Command) {
 	create.Flags().DurationVar(&createOptions.FilteredReplicationWaitTime, "filtered-replication-wait-time", 30*time.Second, "Specifies the maximum time to wait, in seconds, for replication to catch up when syncing tablet streams.")
 	create.Flags().Uint32Var(&createOptions.Limit, "limit", math.MaxUint32, "Max rows to stop comparing after.")
 	create.Flags().BoolVar(&createOptions.DebugQuery, "debug-query", false, "Adds a mysql query to the report that can be used for further debugging.")
-	create.Flags().Uint64Var(&createOptions.MaxReportSampleRows, "max-report-sample-rows", 10, "Maximum number of row differences to report (0 for all differences). NOTE: when increasing this value it is highly recommended to also specify --only-pks")
+	create.Flags().Uint32Var(&createOptions.MaxReportSampleRows, "max-report-sample-rows", 10, "Maximum number of row differences to report (0 for all differences). NOTE: when increasing this value it is highly recommended to also specify --only-pks")
 	create.Flags().BoolVar(&createOptions.OnlyPKs, "only-pks", false, "When reporting missing rows, only show primary keys in the report.")
 	create.Flags().StringSliceVar(&createOptions.Tables, "tables", nil, "Only run vdiff for these tables in the workflow.")
 	create.Flags().Uint32Var(&createOptions.MaxExtraRowsToCompare, "max-extra-rows-to-compare", 1000, "If there are collation differences between the source and target, you can have rows that are identical but simply returned in a different order from MySQL. We will do a second pass to compare the rows for any actual differences in this case and this flag allows you to control the resources used for this operation.")
